ml-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
xgboost==2.1.3
pandas==2.2.3
numpy==2.2.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
nba_api==1.5.2
